clean up stubs
