--- cupy_backends/cuda/stub/cupy_cusparse.h
@@ -1,47 +1,0 @@
-// This file is a stub header file of cusparse for Read the Docs.
-
-#ifndef INCLUDE_GUARD_STUB_CUPY_CUSPARSE_H
-#define INCLUDE_GUARD_STUB_CUPY_CUSPARSE_H
-
-///////////////////////////////////////////////////////////////////////////////
-// Definitions are for compatibility
-///////////////////////////////////////////////////////////////////////////////
-
-extern "C" {
-
-cusparseStatus_t cusparseSnnz_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-cusparseStatus_t cusparseDnnz_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-cusparseStatus_t cusparseCnnz_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-cusparseStatus_t cusparseZnnz_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-cusparseStatus_t cusparseScsr2csr_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-cusparseStatus_t cusparseDcsr2csr_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-cusparseStatus_t cusparseCcsr2csr_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-cusparseStatus_t cusparseZcsr2csr_compress(...) {
-  return CUSPARSE_STATUS_SUCCESS;
-}
-
-}  // extern "C"
-
-
-#endif  // INCLUDE_GUARD_STUB_CUPY_CUSPARSE_H
