+++ cupy_backends/cupy_cusolver.h
@@ -1,0 +1,17 @@
+#ifndef INCLUDE_GUARD_CUPY_CUSOLVER_H
+#define INCLUDE_GUARD_CUPY_CUSOLVER_H
+
+#if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
+
+#include "cuda/cupy_cusolver.h"
+
+#elif defined(CUPY_USE_HIP) // #if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
+
+#include "hip/cupy_rocsolver.h"
+
+#else // #if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
+
+#include "stub/cupy_cusolver.h"
+
+#endif // #if !defined(CUPY_NO_CUDA) && !defined(CUPY_USE_HIP)
+#endif // #ifndef INCLUDE_GUARD_CUPY_CUSOLVER_H
