+++ cupy_backends/cupy_cutensor.h
@@ -1,12 +1,18 @@
 #ifndef INCLUDE_GUARD_CUPY_CUTENSOR_H
 #define INCLUDE_GUARD_CUPY_CUTENSOR_H
 
-#ifndef CUPY_NO_CUDA
+#ifdef CUPY_USE_HIP
+
+// Since ROCm/HIP does not have cuTENSOR, we simply include the stubs here
+// to avoid code dup.
+#include "stub/cupy_cutensor.h"
+
+#elif !defined(CUPY_NO_CUDA)
 
 #include <library_types.h>
 #include <cutensor.h>
 
-#else // #ifndef CUPY_NO_CUDA
+#else
 
 #include "stub/cupy_cutensor.h"
