+++ cupy_backends/hip/cupy_common.h
