+++ cupy_backends/hip/cupy_cuComplex.h
@@ -1,0 +1,20 @@
+#ifndef INCLUDE_GUARD_HIP_CUPY_COMPLEX_H
+#define INCLUDE_GUARD_HIP_CUPY_COMPLEX_H
+
+extern "C" {
+
+///////////////////////////////////////////////////////////////////////////////
+// cuComplex.h
+///////////////////////////////////////////////////////////////////////////////
+
+struct cuComplex{
+    float x, y;
+};
+
+struct cuDoubleComplex{
+    double x, y;
+};
+
+} // extern "C"
+
+#endif // #ifndef INCLUDE_GUARD_HIP_CUPY_COMPLEX_H
