+++ cupy_backends/hip/cupy_hiprand.h
