+++ cupy_backends/hip/cupy_runtime.h
@@ -6,6 +6,8 @@
 
 extern "C" {
 
+bool hip_environment = true;
+
 // Error handling
 const char* cudaGetErrorName(cudaError_t hipError) {
     return hipGetErrorName(hipError);
