+++ cupy_backends/stub/cupy_cuda.h
@@ -3,7 +3,7 @@
 #ifndef INCLUDE_GUARD_STUB_CUPY_CUDA_H
 #define INCLUDE_GUARD_STUB_CUPY_CUDA_H
 
-#include "../cupy_cuda_common.h"
+#include "cupy_cuda_common.h"
 
 extern "C" {
