+++ cupy_backends/stub/cupy_cuda_runtime.h
@@ -3,9 +3,11 @@
 #ifndef INCLUDE_GUARD_STUB_CUPY_CUDA_RUNTIME_H
 #define INCLUDE_GUARD_STUB_CUPY_CUDA_RUNTIME_H
 
-#include "../cupy_cuda_common.h"
+#include "cupy_cuda_common.h"
 
 extern "C" {
+
+bool hip_environment = false;
 
 // Error handling
 const char* cudaGetErrorName(...) {
