+++ cupy_backends/stub/cupy_curand.h
@@ -1,9 +1,9 @@
 // This file is a stub header file of cuda for Read the Docs.
 
-#ifndef INCLUDE_GUARD_SUTB_CUPY_CURAND_H
-#define INCLUDE_GUARD_SUTB_CUPY_CURAND_H
+#ifndef INCLUDE_GUARD_STUB_CUPY_CURAND_H
+#define INCLUDE_GUARD_STUB_CUPY_CURAND_H
 
-#include "../cupy_cuda_common.h"
+#include "cupy_cuda_common.h"
 
 extern "C" {
 
@@ -91,4 +91,4 @@ curandStatus_t curandGeneratePoisson(...) {
 
 } // extern "C"
 
-#endif // #ifndef INCLUDE_GUARD_SUTB_CUPY_CURAND_H
+#endif // #ifndef INCLUDE_GUARD_STUB_CUPY_CURAND_H
